vacation_planner/requirements.txt
langgraph==0.0.38
llama-index==0.10.20
streamlit==1.29.0
google-generativeai==0.4.0
requests==2.31.0
python-dotenv==1.0.0
pandas==2.1.1
plotly==5.18.0
folium==0.14.0
